aufs: copy-up 7/7, white-out
Imagine a sequence such like this.
- user opens a file which exists on the lower readonly branch.
- user unlinks the file (still opened). its parent dir may be removed
  too.
- user writes something to the file.

Then aufs has to copy-up the unlinked file.
Note that the copy-up in aufs is not done in open(2) by default.
This commit handles this case by copy-up the file in write(2).
The target directory is the one which aufs created internally for this
purpose.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>