for aufs hnotify and dcsub: make d_walk() global
When the notify is fired (see the succeeding commits), aufs tries
identifying the changed dentry from dcache. VFS internal function
d_walk() is necessary to do this. I can guess people will say this
function should not be open to filesystems, but I don't have another
idea.
Note that its declaration is not made public. I know this is not good
either.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>