aufs: dirren 5/6, lookup and revalidate with loading the rename info
When aufs meets a new dir inode on a branch in lookup, it tests whether
the inode is in the list which the branch has. If the inode is found, it
means the dir has ever been logically renamed and there is some info
about the name under that dir. Then aufs tries loading the info, and
continues looking up using the before-renamed name on the lower
branches.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>