[vpsAdminOS] nfsd: add export option root_uid
This patch introduces a new NFS export option root_uid, which can be
used to specify what uid should be considered as privileged by the NFS
server in addition to the global root. This is useful when exporting
ZFS filesystems with uid/gid map set. These exports are mounted in the
init_user_ns, so NFS does not perform uid/gid translation and we need
to make selected unprivileged user a root. The export can be configured
e.g. like:

  /directory  <host>(rw,no_root_squash,no_subtree_check,root_uid=100000)

where the namespaced root maps to 100000 on the host and remote NFS
servers.