nfsd: treat userns roots from selected ranges as a real root
We're exporting ZFS datasets with uidmap/gidmap set. When these exports
are mounted and passed on to containers with the respective user
namespace mapping, the container's root user is not recognized and is
treated as a common unprivileged user. So he cannot access and
manipulate files which he does not have explicit permission on.

This patch makes it so that unprivileged users that are roots within
user namespaces of a specific configuration we're using at vpsFree.cz
are recognized as real root. The first such user has id 131072, then all
increments by 524288, which is how many IDs our containers can use.