nfsd: add export option root_uid
We're exporting ZFS datasets with uidmap/gidmap set. When these exports
are mounted and passed on to containers with the respective user
namespace mapping, the container's root user is not recognized and is
treated as a common unprivileged user. So he cannot access and
manipulate files which he does not have explicit permission on.

This patch introduces a new export option root_uid. It can be used to
specify which user should be recognized as root, in addition to the
global root.