aufs: minor, for userspace, don't use typeof
This is detected by the commit in v5.3-rc1,
	d6fc9fcbaa65 2019-07-08 kbuild: compile-test exported headers to
		ensure they are self-contained
But the issue is not for v5.3-rc1 only I am afraid.  So this commit is
brought to aufs4.14 which is my current development base.

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>
(cherry picked from commit d38615f7fef1c65e30e3cc418daa9e19c93ed98f)